refactor: move FormalMuxTester
